# internal/infra/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.18.0

package db

import ()

type Gym struct {
	ID       string
	UserID   string
	GymName  string
	TeamName string
	Active   bool
}

type Student struct {
	ID           string
	GymID        string
	Name         string
	Graduation   string
	Active       bool
	TrainingTime string
}

type User struct {
	ID       string
	Username string
	Password string `json:"-"`
	Email    string
	RoleID   string
	Active   bool
}

type UserRole struct {
	ID   string
	Name string
}